requirements.txt
numpy
scikit-learn
pillow
torch
torchvision
pandas
